refactor: Move schedule loop into Schedule
